% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak.R
\name{peak}
\alias{peak}
\title{Create peak}
\usage{
peak(v, v0, fwhh, p, a, b)
}
\arguments{
\item{v}{WIP}

\item{v0}{WIP}

\item{fwhh}{WIP}

\item{p}{WIP}

\item{a}{WIP}

\item{b}{WIP}
}
\value{
WIP
}
\description{
Create a peak with shape ....
}
\author{
Aswin Verhoeven
}
