% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfit.R
\name{peakfit}
\alias{peakfit}
\title{Do a peak fit}
\usage{
peakfit(v, v0start, fwhhstart, pstart, astart, bstart, sp)
}
\arguments{
\item{v}{WIP}

\item{v0start}{WIP}

\item{fwhhstart}{WIP}

\item{pstart}{WIP}

\item{astart}{WIP}

\item{bstart}{WIP}

\item{sp}{WIP}
}
\value{
WIP
}
\description{
Do peak fit
}
\author{
Aswin Verhoeven
}
