% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakfitsimple.R
\name{peakfitsimple}
\alias{peakfitsimple}
\title{Do a simple peak fit}
\usage{
peakfitsimple(v, v0start, fwhhstart, pstart, astart, bstart, sp)
}
\arguments{
\item{v}{WIP}

\item{v0start}{WIP}

\item{fwhhstart}{WIP}

\item{pstart}{WIP}

\item{astart}{WIP}

\item{bstart}{WIP}

\item{sp}{WIP}
}
\value{
WIP
}
\description{
Do simple peak fit
}
\author{
Aswin Verhoeven
}
